Move all non-moved tests to e2e suite and remove old test suite
